Enable dash UI for CM2 devices
