+++ Resources/Strings.de.resx
@@ -366,6 +366,7 @@
   <data name="Toggle_On" xml:space="preserve"><value>AN</value></data>
   <data name="TabHeader_Inputs" xml:space="preserve"><value>Eingaben</value></data>
   <data name="TabHeader_Dashboard" xml:space="preserve"><value>Dashboard</value></data>
+  <data name="TabHeader_Leds" xml:space="preserve"><value>LEDs</value></data>
   <data name="TabHeader_Rpm" xml:space="preserve"><value>RPM</value></data>
   <data name="TabHeader_Buttons" xml:space="preserve"><value>Tasten</value></data>
   <data name="TabHeader_Knobs" xml:space="preserve"><value>Drehregler</value></data>
