+++ Resources/Strings.es.resx
@@ -371,6 +371,7 @@
   <data name="Toggle_On" xml:space="preserve"><value>ON</value></data>
   <data name="TabHeader_Inputs" xml:space="preserve"><value>Entradas</value></data>
   <data name="TabHeader_Dashboard" xml:space="preserve"><value>Tablero</value></data>
+  <data name="TabHeader_Leds" xml:space="preserve"><value>LEDs</value></data>
   <data name="TabHeader_Rpm" xml:space="preserve"><value>RPM</value></data>
   <data name="TabHeader_Buttons" xml:space="preserve"><value>Botones</value></data>
   <data name="TabHeader_Knobs" xml:space="preserve"><value>Perillas</value></data>
