+++ Resources/Strings.fr.resx
@@ -366,6 +366,7 @@
   <data name="Toggle_On" xml:space="preserve"><value>ON</value></data>
   <data name="TabHeader_Inputs" xml:space="preserve"><value>Entrées</value></data>
   <data name="TabHeader_Dashboard" xml:space="preserve"><value>Tableau de bord</value></data>
+  <data name="TabHeader_Leds" xml:space="preserve"><value>LEDs</value></data>
   <data name="TabHeader_Rpm" xml:space="preserve"><value>RPM</value></data>
   <data name="TabHeader_Buttons" xml:space="preserve"><value>Boutons</value></data>
   <data name="TabHeader_Knobs" xml:space="preserve"><value>Encodeurs</value></data>
