+++ Resources/Strings.it.resx
@@ -369,6 +369,7 @@
   <data name="Toggle_On" xml:space="preserve"><value>ON</value></data>
   <data name="TabHeader_Inputs" xml:space="preserve"><value>Ingressi</value></data>
   <data name="TabHeader_Dashboard" xml:space="preserve"><value>Dashboard</value></data>
+  <data name="TabHeader_Leds" xml:space="preserve"><value>LEDs</value></data>
   <data name="TabHeader_Rpm" xml:space="preserve"><value>RPM</value></data>
   <data name="TabHeader_Buttons" xml:space="preserve"><value>Pulsanti</value></data>
   <data name="TabHeader_Knobs" xml:space="preserve"><value>Manopole</value></data>
