+++ Resources/Strings.zh-Hans.resx
@@ -366,6 +366,7 @@
   <data name="Toggle_On" xml:space="preserve"><value>开</value></data>
   <data name="TabHeader_Inputs" xml:space="preserve"><value>输入</value></data>
   <data name="TabHeader_Dashboard" xml:space="preserve"><value>仪表盘</value></data>
+  <data name="TabHeader_Leds" xml:space="preserve"><value>LEDs</value></data>
   <data name="TabHeader_Rpm" xml:space="preserve"><value>RPM</value></data>
   <data name="TabHeader_Buttons" xml:space="preserve"><value>按键</value></data>
   <data name="TabHeader_Knobs" xml:space="preserve"><value>旋钮</value></data>
